I have made it so that the weeks pregnant for a current pregnancy are calculated fro mthe lmp instead of entered by the user.

+++ iTrust/src/main/edu/ncsu/csc/itrust/model/obstetrics/ObstetricsValidator.java
@@ -67,9 +67,9 @@ public class ObstetricsValidator extends POJOValidator<ObstetricsPregnancy> {
 		}
 		
 		if ( !op.getWeeksPregnant().equals( "" ) ) {
-			int weeksPreg = 0;
+			double weeksPreg = 0.0;
 			try {
-				weeksPreg = Integer.parseInt( op.getWeeksPregnant() );
+				weeksPreg = Double.parseDouble( op.getWeeksPregnant() );
 			} catch ( Exception e ) {
 				errorList.addIfNotNull( "Weeks Pregnant must be numeric" );
 			}
